requirements.txt
Django>=3.0.6,<3.1
uWSGI>=2.0.18,<2.1
djangorestframework>=3.12,<3.13
djangorestframework-jwt>=1.11,<1.12
drf-yasg2>=1.19,<1.20
psycopg2-binary>=2.8,<2.9
packaging>=20.4,<20.5
Pillow>=8.0,<8.1
django-mptt>=0.11

